refactor: move buildSearchBarLayout to src/data (refs #191)
